requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohappyeyeballs==2.4.3
aiohttp==3.9.0
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.30.0
attrs==23.2.0
automium==0.2.6
automium-web==0.1.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.115.0
Flask==3.1.0
frozenlist==1.4.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gspread==6.1.4
h11==0.14.0
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
legacy==0.1.7
magic-filter==1.0.12
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.0.5
oauth2client==4.1.3
oauthlib==3.2.2
propcache==0.2.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.7.0
pydantic-settings==2.7.0
pydantic_core==2.18.1
pyparsing==3.2.1
pystyle==2.9
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.6
telebot==0.0.5
typing_extensions==4.12.2
urllib3==2.2.2
uuid==1.30
uvicorn==0.31.1
Werkzeug==3.1.3
yarl==1.9.4
